+++ src/render/pipelines.ts
@@ -11,12 +11,8 @@ import {
 import { createCyStruct, CyToTS } from "./gpu-struct.js";
 import { MeshHandle } from "./mesh-pool.js";
 import { getAABBFromMesh, Mesh } from "./mesh.js";
-import {
-  cloth_shader,
-  mesh_shader,
-  particle_shader,
-  rope_shader,
-} from "./shaders.js";
+import { cloth_shader, particle_shader, rope_shader } from "./shaders.js";
+import { sceneBufPtr, canvasDepthTex } from "./std-pipeline.js";
 
 // TODO:
 //  [x] pipeline attachements / outputs
@@ -38,82 +34,12 @@ import {
 //  [ ] multiple pipeline outputs
 //  [ ] deferred rendering
 
-// TODO(@darzu):
-export const VertexStruct = createCyStruct(
-  {
-    position: "vec3<f32>",
-    color: "vec3<f32>",
-    normal: "vec3<f32>",
-    uv: "vec2<f32>",
-  },
-  {
-    isCompact: true,
-    serializer: ({ position, color, normal, uv }, _, offsets_32, views) => {
-      views.f32.set(position, offsets_32[0]);
-      views.f32.set(color, offsets_32[1]);
-      views.f32.set(normal, offsets_32[2]);
-      views.f32.set(uv, offsets_32[3]);
-    },
-  }
-);
-export type VertexTS = CyToTS<typeof VertexStruct.desc>;
-
 export const RopeStickStruct = createCyStruct({
   aIdx: "u32",
   bIdx: "u32",
   length: "f32",
 });
 export type RopeStickTS = CyToTS<typeof RopeStickStruct.desc>;
-
-export const SceneStruct = createCyStruct(
-  {
-    cameraViewProjMatrix: "mat4x4<f32>",
-    light1Dir: "vec3<f32>",
-    light2Dir: "vec3<f32>",
-    light3Dir: "vec3<f32>",
-    cameraPos: "vec3<f32>",
-    playerPos: "vec2<f32>",
-    time: "f32",
-  },
-  {
-    isUniform: true,
-    serializer: (data, _, offsets_32, views) => {
-      views.f32.set(data.cameraViewProjMatrix, offsets_32[0]);
-      views.f32.set(data.light1Dir, offsets_32[1]);
-      views.f32.set(data.light2Dir, offsets_32[2]);
-      views.f32.set(data.light3Dir, offsets_32[3]);
-      views.f32.set(data.cameraPos, offsets_32[4]);
-      views.f32.set(data.playerPos, offsets_32[5]);
-      views.f32[offsets_32[6]] = data.time;
-    },
-  }
-);
-export type SceneTS = CyToTS<typeof SceneStruct.desc>;
-
-export function setupScene(): SceneTS {
-  // create a directional light and compute it's projection (for shadows) and direction
-  const worldOrigin = vec3.fromValues(0, 0, 0);
-  const D = 50;
-  const light1Pos = vec3.fromValues(D, D * 2, D);
-  const light2Pos = vec3.fromValues(-D, D * 1, D);
-  const light3Pos = vec3.fromValues(0, D * 0.5, -D);
-  const light1Dir = vec3.subtract(vec3.create(), worldOrigin, light1Pos);
-  vec3.normalize(light1Dir, light1Dir);
-  const light2Dir = vec3.subtract(vec3.create(), worldOrigin, light2Pos);
-  vec3.normalize(light2Dir, light2Dir);
-  const light3Dir = vec3.subtract(vec3.create(), worldOrigin, light3Pos);
-  vec3.normalize(light3Dir, light3Dir);
-
-  return {
-    cameraViewProjMatrix: mat4.create(), // updated later
-    light1Dir,
-    light2Dir,
-    light3Dir,
-    cameraPos: vec3.create(), // updated later
-    playerPos: [0, 0], // updated later
-    time: 0, // updated later
-  };
-}
 
 export const RopePointStruct = createCyStruct(
   {
@@ -131,28 +57,6 @@ export const RopePointStruct = createCyStruct(
   }
 );
 export type RopePointTS = CyToTS<typeof RopePointStruct.desc>;
-
-export const MeshUniformStruct = createCyStruct(
-  {
-    transform: "mat4x4<f32>",
-    aabbMin: "vec3<f32>",
-    aabbMax: "vec3<f32>",
-    tint: "vec3<f32>",
-  },
-  {
-    isUniform: true,
-    serializer: (d, _, offsets_32, views) => {
-      views.f32.set(d.transform, offsets_32[0]);
-      views.f32.set(d.aabbMin, offsets_32[1]);
-      views.f32.set(d.aabbMax, offsets_32[2]);
-      views.f32.set(d.tint, offsets_32[3]);
-    },
-  }
-);
-export type MeshUniformTS = CyToTS<typeof MeshUniformStruct.desc>;
-
-// TODO(@darzu): IMPL
-export type MeshHandleStd = MeshHandle<typeof MeshUniformStruct.desc>;
 
 export const CLOTH_W = 12;
 
@@ -238,10 +142,6 @@ const genRopeStickData = () => {
   return _initRopeStickData;
 };
 
-const sceneBufPtr = CY.registerOneBufPtr("scene", {
-  struct: SceneStruct,
-  init: setupScene,
-});
 const ropePointBufPtr = CY.registerManyBufPtr("ropePoint", {
   struct: RopePointStruct,
   init: genRopePointData,
@@ -285,13 +185,6 @@ const particleIdxBufPtr = CY.registerIdxBufPtr("particleIdx", {
   init: initParticleIdxData,
 });
 
-const canvasDepthTex = CY.registerDepthTexPtr("canvasDepth", {
-  size: [100, 100],
-  onCanvasResize: (w, h) => [w, h],
-  format: "depth24plus-stencil8",
-  init: () => undefined,
-});
-
 const renderRopePipelineDesc = CY.registerRenderPipeline("renderRope", {
   resources: [sceneBufPtr],
   meshOpt: {
@@ -330,85 +223,6 @@ const clothTexPtr0 = CY.registerTexPtr("clothTex0", {
 });
 const clothTexPtr1 = CY.registerTexPtr("clothTex1", {
   ...clothTexPtrDesc,
-});
-
-export const MAX_MESHES = 20000;
-export const MAX_VERTICES = 21844;
-
-const meshVertsPtr = CY.registerManyBufPtr("meshVertsBuf", {
-  struct: VertexStruct,
-  init: () => MAX_VERTICES,
-});
-
-const meshTriIndsPtr = CY.registerIdxBufPtr("meshTriIndsBuf", {
-  init: () => MAX_VERTICES,
-});
-
-const meshLineIndsPtr = CY.registerIdxBufPtr("meshLineIndsBuf", {
-  init: () => MAX_VERTICES * 2,
-});
-
-const meshUnisPtr = CY.registerManyBufPtr("meshUni", {
-  struct: MeshUniformStruct,
-  init: () => MAX_MESHES,
-});
-
-export const meshPoolPtr = CY.registerMeshPoolPtr("meshPool", {
-  computeVertsData,
-  computeUniData,
-  vertsPtr: meshVertsPtr,
-  unisPtr: meshUnisPtr,
-  triIndsPtr: meshTriIndsPtr,
-  lineIndsPtr: meshLineIndsPtr,
-});
-
-export function computeUniData(m: Mesh): MeshUniformTS {
-  const { min, max } = getAABBFromMesh(m);
-  const uni: MeshUniformTS = {
-    transform: mat4.create(),
-    aabbMin: min,
-    aabbMax: max,
-    tint: vec3.create(),
-  };
-  return uni;
-}
-
-export function computeVertsData(m: Mesh): VertexTS[] {
-  const vertsData: VertexTS[] = m.pos.map((pos, i) => ({
-    position: pos,
-    color: [0.0, 0.0, 0.0],
-    normal: [1.0, 0.0, 0.0],
-    uv: m.uvs ? m.uvs[i] : [0.0, 0.0],
-  }));
-  m.tri.forEach((triInd, i) => {
-    // set provoking vertex data
-    // TODO(@darzu): add support for writting to all three vertices (for non-provoking vertex setups)
-    const normal = computeTriangleNormal(
-      m.pos[triInd[0]],
-      m.pos[triInd[1]],
-      m.pos[triInd[2]]
-    );
-    vertsData[triInd[0]].normal = normal;
-    vertsData[triInd[0]].color = m.colors[i];
-  });
-  return vertsData;
-}
-
-const renderTriPipelineDesc = CY.registerRenderPipeline("triRender", {
-  resources: [
-    sceneBufPtr,
-    // TODO(@darzu): support textures
-    { ptr: clothTexPtr0, access: "read", alias: "clothTex" },
-  ],
-  meshOpt: {
-    pool: meshPoolPtr,
-    stepMode: "per-mesh-handle",
-  },
-  shader: mesh_shader,
-  shaderVertexEntry: "vert_main",
-  shaderFragmentEntry: "frag_main",
-  output: canvasTexturePtr,
-  depthStencil: canvasDepthTex,
 });
 
 // TODO(@darzu): CLOTH
