+++ src/render/shaders.ts
@@ -1,69 +1,5 @@
 import { CLOTH_W } from "./pipelines.js";
 
-export const mesh_shader = () =>
-  `
-  struct VertexOutput {
-      @location(0) @interpolate(flat) normal : vec3<f32>,
-      @location(1) @interpolate(flat) color : vec3<f32>,
-      @location(2) worldPos: vec4<f32>,
-      @builtin(position) position : vec4<f32>,
-  };
-
-  @stage(vertex)
-  fn vert_main(input: VertexInput) -> VertexOutput {
-      let position = input.position;
-      let uv = input.uv;
-      let color = input.color;
-      let normal = input.normal;
-
-      var output : VertexOutput;
-      let worldPos: vec4<f32> = meshUni.transform * vec4<f32>(position, 1.0);
-
-      // let uvInt: vec2<i32> = vec2<i32>(5, 5);
-      // let uvInt: vec2<i32> = vec2<i32>(10, i32(uv.x + 5.0));
-      let uvInt: vec2<i32> = vec2<i32>(i32(uv.x * 10.0), i32(uv.y * 10.0));
-      let texDisp = textureLoad(clothTex, uvInt, 0);
-
-      // let finalPos = worldPos;
-      // TODO: support texture resource
-      // let finalPos = vec4<f32>(worldPos.xy, worldPos.z + uv.x * 10.0, worldPos.w);
-      let finalPos = vec4<f32>(worldPos.xyz + texDisp.xyz, 1.0);
-
-      output.worldPos = finalPos;
-      output.position = scene.cameraViewProjMatrix * finalPos;
-      output.normal = normalize(meshUni.transform * vec4<f32>(normal, 0.0)).xyz;
-      // output.color = vec3<f32>(f32(uvInt.x), f32(uvInt.y), 1.0);
-      // output.color = texDisp.rgb;
-      // output.color = vec3(uv.xy, 1.0);
-      output.color = color + meshUni.tint;
-      return output;
-  }
-
-  @stage(fragment)
-  fn frag_main(input: VertexOutput) -> @location(0) vec4<f32> {
-      let normal = input.normal;
-      // let normal = -normalize(cross(dpdx(input.worldPos.xyz), dpdy(input.worldPos.xyz)));
-      let light1 : f32 = clamp(dot(-scene.light1Dir, normal), 0.0, 1.0);
-      let light2 : f32 = clamp(dot(-scene.light2Dir, normal), 0.0, 1.0);
-      let light3 : f32 = clamp(dot(-scene.light3Dir, normal), 0.0, 1.0);
-      let resultColor: vec3<f32> = input.color 
-        * (light1 * 1.5 + light2 * 0.5 + light3 * 0.2 + 0.1);
-      let gammaCorrected: vec3<f32> = pow(resultColor, vec3<f32>(1.0/2.2));
-
-      let fogDensity: f32 = 0.02;
-      let fogGradient: f32 = 1.5;
-      // let fogDist: f32 = 0.1;
-      let fogDist: f32 = max(-input.worldPos.y - 10.0, 0.0);
-      // output.fogVisibility = 0.9;
-      let fogVisibility: f32 = clamp(exp(-pow(fogDist*fogDensity, fogGradient)), 0.0, 1.0);
-
-      let backgroundColor: vec3<f32> = vec3<f32>(0.6, 0.63, 0.6);
-      let finalColor: vec3<f32> = mix(backgroundColor, gammaCorrected, fogVisibility);
-      return vec4<f32>(finalColor, 1.0);
-      // return vec4<f32>(input.color, 1.0);
-  }
-`;
-
 export const cloth_shader = () =>
   `
 // @group(0) @binding(1) var inTex : texture_2d<f32>;
